Refactored upgrade path handling
